+++ src/jpegxl.ts
@@ -30,6 +30,7 @@ export {
 export {
   convertJpegXlCmykLayerToRgba8,
   convertJpegXlFloat32LayerToRgba16,
+  convertJpegXlFloatLayerToRgba16,
   type JpegXlRgba8Image,
   type JpegXlRgba16Image,
   jpegXlNativeFloat32ColorPlanes,